requirements.txt
streamlit>=1.28.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
pandas>=2.0.0
plotly>=5.17.0 
fastapi>=0.115.0
pydantic>=2.0.0